Улучшение архитектуры: явное разделение логики и отрисовки. Использован паттерн MVC. Tools - промежуточное звено, связывающее логику и отрисовку(паттерн Посредник)
