+++ src/jvmMain/kotlin/application/logic/edge.kt
@@ -1,0 +1,3 @@
+package application.logic
+
+data class Edge(val vertices: Pair<Vertex, Vertex>, var weight: Int = 0)
