+++ src/jvmMain/kotlin/application/logic/logic.kt
@@ -1,0 +1,3 @@
+package application.logic
+
+class Logic(val graph: Graph = Graph())
