+++ src/jvmMain/kotlin/application/logic/vertex.kt
@@ -1,0 +1,3 @@
+package application.logic
+
+data class Vertex(var name: String = "")
