--- src/jvmMain/kotlin/logic/edge.kt
@@ -1,7 +1,0 @@
-package logic
-
-import ui.objects.EdgeUI
-
-class Edge(val vertices: Pair<Vertex, Vertex>, val weight: Int = 0) {
-    val edgeUI = EdgeUI(this)
-}
