--- src/jvmMain/kotlin/logic/vertex.kt
@@ -1,7 +1,0 @@
-package logic
-
-import ui.objects.VertexUI
-
-class Vertex() {
-    val vertexUI = VertexUI(this)
-}
